Set quickfix-title first before adding items
Currently the quickfix title isn't set yet when the 'QuickFixCmdPost'
event is triggered. 'QuickFixCmdPost' is executed after the items have
been added to the list using ':cfile' and friends.

Therefore, we first create a new quickfix list with a given title, and
then add the items to the list.

Having the quickfix title available is helpful to, for example,
automatically sort the quickfix list when the items are from a ripgrep
search since ripgrep's output is unsorted when run in parallel.

Alternatives considered: Add the items to a new list with ':noautocmd',
and then manually trigger 'QuickFixCmdPost' using ':doautocmd' _after_
the quickfix title has been set. However, this will prevent triggering
necessary autocommands when Vim jumps to the first error (after ':cfile'
or ':lfile'). For example, the filetype won't be set if the buffer
hasn't been loaded previously.